requirements.txt
python-dotenv~=1.0.0
fastapi~=0.108.0
aiofiles~=23.2.1
starlette~=0.32.0.post1
pydantic~=2.5.3
passlib~=1.7.4
tortoise-orm
tortoise-orm[asyncpg]
aerich
avatar-generator
python-jose[cryptography]
python-multipart
uvicorn~=0.25.0
pydantic-settings
yarl~=1.9.4
sentry-sdk
loguru~=0.7.2
ujson
aiohttp~=3.9.3
pydantic[email]
DateTime~=5.4
Jinja2~=3.1.3
requests~=2.31.0
pytest~=7.4.4
httpx~=0.26.0
APScheduler~=3.10.4
aiogram~=3.4.1
pytz~=2023.3.post1
transliterate~=1.10.2
yagmail~=0.15.293
redis
